Move CandyStore class into candystore module
